+++ lib/chef-acceptance/cli.rb
@@ -1,13 +1,10 @@
 require 'thor'
-require 'mixlib/shellout'
 require 'chef-acceptance/version'
+require 'chef-acceptance/test_suite'
+require 'chef-acceptance/executable_helper'
 
 module ChefAcceptance
   class Cli < Thor
-    ACCEPTANCE_COOKBOOK_NAME = 'acceptance-cookbook'.freeze
-
-    CORE_ACCEPTANCE_COMMANDS = %w(destroy provision verify).freeze
-
     DESTROY_OPTIONS = %w(always never passing).freeze
 
     package_name 'chef-acceptance'
@@ -15,26 +12,17 @@ module ChefAcceptance
     #
     # Create core acceptance commands
     #
-    CORE_ACCEPTANCE_COMMANDS.each do |command|
-      desc "#{command} TEST_SUITE", "Run #{command}"
-      define_method(command) do |test_suite|
-        chef_client_installed!
+    TestSuite::CORE_ACCEPTANCE_RECIPES.each do |recipe|
+      desc "#{recipe} TEST_SUITE", "Run #{recipe}"
+      define_method(recipe) do |test_suite_name|
+        test_suite = TestSuite.new(test_suite_name)
+        test_suite.run_recipes = [recipe]
 
-        suite = TestSuite.new(test_suite)
-
-        suite.exist!
-
-        shellout = []
-        shellout << 'chef-client -z'
-        shellout << "-c #{File.expand_path(suite.chef_config_file)}"
-        shellout << "-o #{ACCEPTANCE_COOKBOOK_NAME}::#{command}"
-
-        chef_zero = Mixlib::ShellOut.new(shellout.join(' '),
-                                         cwd: suite.acceptance_cookbook_dir,
-                                         live_stream: $stdout)
-        chef_zero.run_command
-
-        abort "#{chef_zero.stdout}\n#{chef_zero.stderr}" if chef_zero.error?
+        begin
+          test_suite.run
+        rescue => e
+          abort e.message
+        end
       end
     end
 
@@ -44,45 +32,44 @@ module ChefAcceptance
            aliases: '-d',
            default: 'passing',
            desc: "Destroy strategy to use after testing (#{DESTROY_OPTIONS.join(', ')})"
-    def test(test_suite)
+    def test(test_suite_name)
       unless DESTROY_OPTIONS.include? options[:destroy]
         abort "destroy option must be one of: #{DESTROY_OPTIONS.join(', ')}"
       end
 
+      test_suite = TestSuite.new(test_suite_name)
+
+      recipe_list = %w(provision verify)
+      recipe_list << 'destroy' if destroy?
+
+      test_suite.run_recipes = recipe_list
       begin
-        provision(test_suite)
-        verify(test_suite)
+        test_suite.run
       rescue => e
-        puts e
-      ensure
-        destroy(test_suite) unless options[:destroy] == 'never'
+        puts e.message
+
+        if destroy?
+          test_suite.run_recipes = ['destroy']
+          test_suite.run
+        end
+
+        abort
       end
     end
 
     desc 'generate NAME', 'Generate acceptance scaffolding'
-    def generate(name)
-      suite = TestSuite.new(name)
+    def generate(test_suite_name)
+      test_suite = TestSuite.new(test_suite_name)
 
-      abort "Test suite '#{name}' already exists." if suite.exist?
+      abort "Test suite '#{test_suite_name}' already exists." if test_suite.exist?
 
-      [suite.acceptance_cookbook_dir, suite.chef_dir, suite.recipes_dir].each do |path|
-        FileUtils.mkpath path
+      begin
+        test_suite.generate
+      rescue => e
+        abort e.message
       end
 
-      CORE_ACCEPTANCE_COMMANDS.each do |file|
-        FileUtils.touch(File.join(suite.recipes_dir, "#{file}.rb"))
-      end
-
-      File.write(File.join(suite.acceptance_cookbook_dir, 'metadata.rb'),
-        "name '#{ACCEPTANCE_COOKBOOK_NAME}'")
-
-      File.write(suite.chef_config_file,
-        "cookbook_path File.join(File.dirname(__FILE__), '..', '..')")
-
-      File.write(File.join(name, '.gitignore'),
-        "local-mode-cache/\nnodes/")
-
-      puts "Run `chef-acceptance test #{name}`"
+      puts "Run `chef-acceptance test #{test_suite_name}`"
     end
 
     desc 'version', 'Print chef-acceptance version'
@@ -93,59 +80,13 @@ module ChefAcceptance
     desc 'info', 'Print chef-acceptance information'
     def info
       puts "chef-acceptance version: #{ChefAcceptance::VERSION}"
-      client = executable_installed? 'chef-client'
+      client = ExecutableHelper.executable_installed? 'chef-client'
       puts "chef-client path: #{client ? client : "not found in #{ENV['PATH']}"}"
     end
 
     no_commands do
-      def executable_installed?(executable)
-        exts = ENV['PATHEXT'] ? ENV['PATHEXT'].split(';') : ['']
-        ENV['PATH'].split(File::PATH_SEPARATOR).each do |path|
-          exts.each do |ext|
-            exe = File.join(path, "#{executable}#{ext}")
-            return exe if File.executable?(exe) && !File.directory?(exe)
-          end
-        end
-        false
-      end
-
-      def chef_client_installed!
-        unless executable_installed? 'chef-client'
-          abort "Could not find chef-client in #{ENV['PATH']}"
-        end
-      end
-
-      TestSuite = Struct.new(:name) do
-        # TODO: sanitize input
-
-        def exist?
-          File.exist?(name)
-        end
-
-        def exist!
-          unless self.exist?
-            abort <<-EOS
-  Could not find test suite '#{name}' in the current working \
-  directory '#{Dir.pwd}'.
-  EOS
-          end
-        end
-
-        def acceptance_cookbook_dir
-          File.join(name, '.acceptance', ACCEPTANCE_COOKBOOK_NAME)
-        end
-
-        def recipes_dir
-          File.join(acceptance_cookbook_dir, 'recipes')
-        end
-
-        def chef_dir
-          File.join(acceptance_cookbook_dir, '.chef')
-        end
-
-        def chef_config_file
-          File.join(chef_dir, 'config.rb')
-        end
+      def destroy?
+        options[:destroy] != 'never'
       end
     end
   end